refactor: rename parse.lua to process.lua
